### 2025-08-14 #8815 @editorialbot
[REVIEW]: PyDTS: A Python Package for Discrete-Time Survival Analysis with Competing Risks and Optional Penalization
**Submitting author:** @tomer1812 (<a href="http://orcid.org/0000-0002-3208-8460">Tomer Meir</a>)
**Repository:** https://github.com/tomer1812/pydts
**Branch with paper.md** (empty if default branch):
**Version:** v0.9.7
