### 2025-08-24 4ad621a
joss paper - review updates
