+++ paper.md
@@ -48,25 +48,14 @@ Moreover, in the continuous-time setting, competing-risks data can often be anal
 To the best of our knowledge, *PyDTS* is the first open-source Python package dedicated for discrete-time survival analysis with competing risks.
 Details on the statistical models and methods implemented in *PyDTS* are summarized in the [package documentation](https://tomer1812.github.io/pydts/) and described in great detail in @meir_gorfine_dtsp_2025.
 
-# Package Details 
 
-## Dependencies
+# Key Features
 
 *PyDTS* can be easily installed via *PyPI* as follows:
 
 `pip install pydts`
 
-This command automatically installs the following dependencies:
-
-- NumPy [@Harris2020] (<2.0.0)
-- pandas [@McKinney2010]
-- lifelines [@Davidson-Pilon:2019]
-- SciPy [@Virtanen2020]
-- scikit-learn [@Pedregosa2011]
-- statsmodels [@Seabold2010]
-
-
-## Key Features
+It includes the following key features:
 
 1. **Estimation procedures:** Two methods are implemented, `TwoStagesFitter` of @meir_gorfine_dtsp_2025, and `DataExpansionFitter` of @lee_analysis_2018. The `TwoStagesFitter` supports both regularization and the inclusion of time-dependent covariates, features that are not available in the `DataExpansionFitter` implementation.
 2. **Sure Independence Screening:** The `SISTwoStagesFitter` class implements the Sure Independence Screening (SIS) of @zhao2012principled. SIS is a powerful dimensionality reduction technique designed for ultra-high-dimensional settings, where the number of covariates far exceeds the number of observations, a situation often encountered in genomic studies and other high-throughput domains. It works by filtering out a large number of uninformative covariates based on their marginal association with the outcome. After screening, penalized variable selection methods (e.g., LASSO) are typically applied to the reduced set of covariates to perform more refined modeling and selection.
@@ -79,11 +68,13 @@ This command automatically installs the following dependencies:
 The utility of *PyDTS* is demonstrated through an analysis of patients' length of stay (LOS) in intensive care unit (ICU), conducted by @meir_gorfine_dtsp_2025. This analysis uses the publicly accessible, large-scale Medical Information Mart for Intensive Care (MIMIC-IV, version 2.0) dataset [@johnson_mimic-iv_2022; @goldberger_physiobank_2000]. 
 
 @meir_gorfine_dtsp_2025 developed a discrete-time survival model to predict ICU LOS based on patients’ clinical characteristics at admission.
-The data comprises 25,170 ICU admissions, with LOS recorded in discrete units from 1 to 28 days, resulting in many patients sharing the same event time on each day. Three competing events are considered: discharge to home (69.0%), transfer to another medical facility (21.4%), and in-hospital death (6.1%). Patients who left the ICU against medical advice (1.0%) are treated as right-censored, and administrative censoring is applied to those hospitalized for more than 28 days (2.5%). The analysis includes 36 covariates per patient. For a full description of the data, see @meir_gorfine_dtsp_2025.
[…]
