### 2025-08-25 9c04319
docs - joss review updates
