### 2025-08-25 252bb15
docs - joss review updates
