+++ README.md
@@ -31,7 +31,7 @@ pip install pydts
 
 ## Quick Start
 
-The following example demonstrates how to generate synthetic data and fit a TwoStagesFitter model.
+The following example demonstrates how to generate synthetic data and fit a `TwoStagesFitter` model.
 
 Detailed definitions and explanations are available in the [methods section](https://tomer1812.github.io/pydts/methods/) of the documentation. 
 
@@ -47,11 +47,11 @@ The function `generate_quick_start_df` simulates a dataset with the following de
   - $\beta_1$ = −log([0.8, 3, 3, 2.5, 2])  
   - $\beta_2$ = −log([1, 3, 4, 3, 2])  
 
-For each patient, a censoring time `C` is drawn from Uniform{1, ..., 14}.
-The observed time is defined as `X = min(T, C)`, where `T` is the event time which is sampled based on the covariates of each patient and the hazard coefficients.
-If censoring occurs before the event (`C < T`), the event type is set to `J = 0`.
+For each patient, a censoring time $C$ is drawn from Uniform{1, ..., 14}.
+The observed time is defined as $X = min(T, C)$, where $T$ is the event time which is sampled based on the covariates of each patient and the hazard coefficients.
+If censoring occurs before the event ($C < T$), the event type is set to $J = 0$.
 
-Once the dataset is generated, you can fit a `TwoStagesFitter` to the data (without columns `C` and `T` which are not observed in practice).
+Once the dataset is generated, you can fit a `TwoStagesFitter` to the data (without columns $C$ and $T$ which are not observed in practice).
 
 You can generate synthetic data and fit your first `TwoStagesFitter` model with the following code: 
 
