### 2025-08-26 bf72d2e
Update paper.md
