+++ docs/index.md
@@ -16,6 +16,7 @@
 - Model evaluation metrics for predictive accuracy and calibration.
 - Simulation tools for generating synthetic datasets for research and testing.
 
+Additional simulations and illustrative examples are available in [Meir and Gorfine (2025), Discrete-Time Competing-Risks Regression with or without Penalization, Biometrics (2025)](https://academic.oup.com/biometrics/article/81/2/ujaf040/8120014), and in the accompanying [Github Repository](https://github.com/tomer1812/DiscreteTimeSurvivalPenalization/tree/main)
 
 ## Installation
 
