### 2025-08-26 #8815 @tomer1812
Hi @rich2355 ,

Thank you so much for all your great comments and suggestions. I have revised the paper, the Github repo, and the documentation accordingly. Specifically:

Paper

1. I removed the Dependencies section as suggested.

2. I rephrased with the following clarification: "The dataset comprises 25,170 ICU patients. For each patient, only the last admission is considered, and features related to prior admission history are included."

3. I added details: "When applying the two-step procedure with LASSO regularization, we need to specify the hyperparameters that control the amount of regularization applied to each competing event. PyDTS provides functionality for tuning these hyperparameters via K-fold cross-validation. By default, the optimal values are those that maximize the out-of-sample global-AUC metric, as defined in Meir and Gorfine (2025), Appendix I. Additional tuning options are also available. Here, a grid search with 4-fold cross-validation was performed to select the optimal hyperparameters that maximize the global-AUC. The code below illustrates such tuning procedure ... "

4. I added:
"Figure 1 presents the results of the selection procedure. Panels A–C illustrate the number of non-zero estimated coefficients, denoted as $\beta_j$, as a function of the regularization hyperparameter $\eta_j$ for each competing event. Panels D–F illustrate the coefficient values as a function of $\eta_j$ for each competing event. Panels G–I illustrate the $\widehat{\mbox{AUC}}_j(t)$ metric for the selected set of $\eta_j$, $j = 1, 2, 3$. A comprehensive description of the case study settings and results can be found in Meir and Gorfine (2025)."

5. I changed it to a pdf file as suggested, do you still see the same distortion? It looks ok to me when I zoom in.

Github Repo

1. I have changed the structure of the README as suggested, removed the figure, and added further details.

2. I expanded the explanations provided in the quick example.

Documentation

1. I removed the figure.

2. I added reminders of the relevant definitions:
"Failure times of observations, denoted as T, were generated based on the model:
...
Censoring time for each observation was sampled from a discrete uniform distribution. The last observed time X is calculated as X=min(C,T), and J is the event-type with J=0 if and only if C<T."

3. I removed the INFO messages.

4. I reviewed the documentation again and corrected the word formatting.

5. Both methods in this comparison are based on their PyDTS implementations. Since Lee et al.’s method was originally implemented in R, this provides Python users with a clearer sense of which estimator is more efficient to use within a Python environment.
I added a clarification to the documentation as well: "We conducted a simulation study demonstrating the performances of Meir et al. (2022) [1] and comparing it with that of Lee et al. (2018) [2]. The PyDTS implementation was used for both methods."

6. The simulations and illustrative example of the Biometrics paper is available in the Biometrics paper Github repo, I added a link to it in the README and to the introduction of the documentation as well:
"Additional simulations and illustrative examples are available in [Meir and Gorfine (2025), Discrete-Time Competing-Risks Regression with or without Penalization, Biometrics (2025)](https://academic.oup.com/biometrics/article/81/2/ujaf040/8120014), and in the accompanying [Github Repository](https://github.com/tomer1812/DiscreteTimeSurvivalPenalization/tree/main)"

Thank you again for all your helpful comments. Please feel free to add further feedback or let me know if any of the corrections require further improvement.

Thank you,
Tomer
