+++ docs/intro.md
@@ -1,17 +1,5 @@
 # Introduction
 
-Based on 
-
-"PyDTS: A Python Package for Discrete-Time Survival Analysis with Competing Risks"
-
-Tomer Meir\*, Rom Gutman\*, and Malka Gorfine (2022) [[1]](#1).
-
-and
-
-"Discrete-time Competing-Risks Regression with or without Penalization"
-
-Tomer Meir and Malka Gorfine (2023) [[2]](#2).
-
 ## Discrete-data survival analysis
 Discrete-data survival analysis refers to the case where data can only take values over a discrete grid. Sometimes, events can only occur at regular, discrete points in time. For example, in the United States a change in party controlling the presidency only occurs quadrennially in the month of January [[3]](#3). In other situations events may occur at any point in time, but available data record only the particular interval of time in which each event occurs. For example, death from cancer measured by months since time of diagnosis [[4]](#4), or length of stay in hospital recorded on a daily basis. It is well-known that naively using standard continuous-time models (even after correcting for ties) with discrete-time data may result in biased estimators for the discrete time models.
 
@@ -20,12 +8,12 @@ Discrete-data survival analysis refers to the case where data can only take valu
 Competing events arise when individuals are susceptible to several types of events but can experience at most one event. For example, competing risks for hospital length of stay are discharge and in-hospital death. Occurrence of one of these events precludes us from observing the other event on this patient. Another classical example of competing risks is cause-specific mortality, such as death from heart disease, death from cancer and death from other causes [[5, 6]](#5#6). 
 
 
-PyDTS is an open source Python package which implements tools for discrete-time survival analysis with competing risks.
+*PyDTS* is an open source Python package which implements tools for discrete-time survival analysis with competing risks.
 
 
 ## References
 <a id="1">[1]</a> 
-Meir, Tomer\*, Gutman, Rom\*, and Gorfine, Malka, 
+Meir, Tomer, Gutman, Rom, and Gorfine, Malka, 
 "PyDTS: A Python Package for Discrete-Time Survival Analysis with Competing Risks"
 (2022)
 
