+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "pydts"
-version = "0.9.7"
+version = "0.9.8"
 description = "Discrete time survival analysis with competing risks"
 authors = ["Tomer Meir <tomer1812@gmail.com>", "Rom Gutman <rom.gutman1@gmail.com>", "Malka Gorfine <malkago12@gmail.com>"]
 license = "GNU GPLv3"
@@ -11,47 +11,65 @@ keywords = ["Discrete Time", "Time to Event" ,"Survival Analysis", "Competing Ev
 documentation = "https://tomer1812.github.io/pydts"
 
 [tool.poetry.dependencies]
-python = ">=3.9,<3.11"
-pandas = "^1.4.1"
-mkdocs = "^1.4.3"
-mkdocs-material = "^9.0.0"
-#mknotebooks = "^0.7.1"
-lifelines = "^0.26.4"
-scipy = "^1.8.0"
-scikit-learn = "^1.0.2"
-tqdm = "^4.63.0"
-statsmodels = "^0.13.2"
-pandarallel = "^1.5.7"
-ipython = "^8.2.0"
-numpy = ">=1.23.4, <2.0.0"
-psutil = "^5.9.4"
-setuptools = "^68.0.0"
-seaborn = "^0.12.2"
-mkdocstrings = "^0.28"
-mkdocstrings-python = "^1.16"
-mknotebooks = "^0.8.0"
+python = ">=3.9,<3.14"
 
-[tool.poetry.dev-dependencies]
-pytest = "^7.0.1"
-coverage = {extras = ["toml"], version = "^6.3.2"}
-pytest-cov = "^3.0.0"
-mkdocs = "^1.2.3"
-mkdocs-material = "^9.0.0"
-#mknotebooks = "^0.7.1"
-jupyter = "^1.0.0"
-scikit-survival = "^0.17.1"
-tableone = "^0.7.10"
+# NumPy varies by Python version
+numpy = [
+  { version = ">=1.26,<2.1", python = ">=3.9,<3.10" },   # Py3.9: allow 1.26–2.0.x
+  { version = ">=1.26,<2.3", python = ">=3.10,<3.11" },  # Py3.10: allow up to 2.2.x
+  { version = ">=1.26",       python = ">=3.11,<3.14" }  # Py3.11–3.13: 1.26+ incl. 2.1–2.3.x
+]
+
+# SciPy varies by Python version
+scipy = [
+  { version = ">=1.13,<1.14", python = ">=3.9,<3.10" },   # Py3.9
+  { version = ">=1.14,<1.16", python = ">=3.10,<3.11" },  # Py3.10 (1.16+ drops 3.10)
+  { version = ">=1.15",       python = ">=3.11,<3.14" }   # Py3.11–3.13
+]
+
+# The rest can be simple lower-bounds
+pandas        = ">=2.2.2"     # first generally compatible with NumPy 2.0
+scikit-learn  = ">=1.6"       # supports Py 3.9–3.13
+statsmodels   = ">=0.14.2"    # NumPy 2 compatibility
+lifelines     = ">=0.27"
+tqdm          = ">=4.66"
+psutil        = ">=5.9"
+seaborn       = ">=0.13"
+
+
+[tool.poetry.group.dev.dependencies]
+pytest = ">=8.4"
+pytest-cov = ">=6.2"
+coverage = { version = ">=7.5,<8", extras = ["toml"] }  # optional but useful
+
+
+[tool.poetry.group.docs.dependencies]
+mkdocs = ">=1.6,<2"
+mkdocs-material = ">=9.6,<10"
+mkdocstrings = ">=0.30"
+mkdocstrings-python = ">=1.18"
[…]
