+++ .github/workflows/tests.yml
@@ -1,151 +1,61 @@
 name: Tests
 on: [push, pull_request]
+
 jobs:
-  test-DataExpansionFitter:
-    runs-on: ubuntu-latest
-    timeout-minutes: 60
-    strategy:
-      matrix:
-        python-version: ['3.10']
-    steps:
-      - uses: actions/checkout@v4
-      - name: Set up Python ${{ matrix.python-version }}
-        uses: actions/setup-python@v4
-        with:
-          python-version: ${{ matrix.python-version }}
-      - name: Setup
-        run: |
-          sudo apt-get update && sudo apt-get install -y gettext
-          pip install poetry
-          poetry install
-      - name: Run tests
-        run: poetry run pytest tests/test_DataExpansionFitter.py -v
-
-  test-EventTimesSampler:
-    runs-on: ubuntu-latest
-    timeout-minutes: 60
-    strategy:
-      matrix:
-        python-version: ['3.10']
-    steps:
-      - uses: actions/checkout@v4
-      - name: Set up Python ${{ matrix.python-version }}
-        uses: actions/setup-python@v4
-        with:
-          python-version: ${{ matrix.python-version }}
-      - name: Setup
-        run: |
-          sudo apt-get update && sudo apt-get install -y gettext
-          pip install poetry
-          poetry install
-      - name: Run tests
-        run: poetry run pytest tests/test_EventTimesSampler.py -v
-
-  test-TwoStagesFitter:
-    runs-on: ubuntu-latest
-    timeout-minutes: 60
-    strategy:
-      matrix:
-        python-version: ['3.10']
-    steps:
-      - uses: actions/checkout@v4
-      - name: Set up Python ${{ matrix.python-version }}
-        uses: actions/setup-python@v4
-        with:
-          python-version: ${{ matrix.python-version }}
-      - name: Setup
-        run: |
-          sudo apt-get update && sudo apt-get install -y gettext
-          pip install poetry
-          poetry install
-      - name: Run tests
-        run: poetry run pytest tests/test_TwoStagesFitter.py -v
-
-  test-TwoStagesFitterExact:
+  tests:
+    name: ${{ matrix.test-name }} (py${{ matrix.python-version }})
     runs-on: ubuntu-latest
     timeout-minutes: 120
     strategy:
+      fail-fast: false
       matrix:
-        python-version: ['3.10']
+        python-version: ["3.9", "3.10", "3.11", "3.12", "3.13"]
+        test-name:
+          - DataExpansionFitter
+          - EventTimesSampler
+          - TwoStagesFitter
+          - TwoStagesFitterExact
+          - screening
+          - model_selection
+          - remaining
+    permissions:
+      contents: read
     steps:
       - uses: actions/checkout@v4
+
       - name: Set up Python ${{ matrix.python-version }}
-        uses: actions/setup-python@v4
+        uses: actions/setup-python@v5
         with:
           python-version: ${{ matrix.python-version }}
-      - name: Setup
-        run: |
-          sudo apt-get update && sudo apt-get install -y gettext
-          pip install poetry
-          poetry install
-      - name: Run tests
-        run: poetry run pytest tests/test_TwoStagesFitterExact.py -v
+          cache: poetry   # caches Poetry virtualenv(s) keyed by poetry.lock
 
-  test-screening:
-    runs-on: ubuntu-latest
-    timeout-minutes: 60
-    strategy:
-      matrix:
-        python-version: ['3.10']
-    steps:
-      - uses: actions/checkout@v4
-      - name: Set up Python ${{ matrix.python-version }}
-        uses: actions/setup-python@v4
-        with:
-          python-version: ${{ matrix.python-version }}
-      - name: Setup
-        run: |
-          sudo apt-get update && sudo apt-get install -y gettext
-          pip install poetry
[…]
