### 2025-08-31 #8815 @tomer1812
Hi @rich2355,

Thank you for letting me know!
I had to restrict some versions due to compatibility problems between dependencies and Python versions. I will work on extending PyDTS's compatibility, or at least updating it, and will let you know once it’s resolved.
I will add the final dependencies to the README as you suggested.

Thanks again,
Tomer
