+++ .github/workflows/tests.yml
@@ -9,33 +9,26 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        python-version: ["3.9", "3.10", "3.11", "3.12", "3.13"]
-        test-name:
-          - DataExpansionFitter
-          - EventTimesSampler
-          - TwoStagesFitter
-          - TwoStagesFitterExact
-          - screening
-          - model_selection
-          - remaining
-    permissions:
-      contents: read
+        python-version: ["3.9","3.10","3.11","3.12","3.13"]
+        test-name: [DataExpansionFitter, EventTimesSampler, TwoStagesFitter,
+                    TwoStagesFitterExact, screening, model_selection, remaining]
     steps:
       - uses: actions/checkout@v4
 
-      - name: Set up Python ${{ matrix.python-version }}
-        uses: actions/setup-python@v5
+      - uses: actions/setup-python@v5
         with:
           python-version: ${{ matrix.python-version }}
-          cache: poetry   # caches Poetry virtualenv(s) keyed by poetry.lock
+          cache: poetry            # cache Poetry deps keyed by poetry.lock
+
+      - uses: snok/install-poetry@v1
+        with:
+          version: latest
+          virtualenvs-create: true
+          virtualenvs-in-project: true
+          installer-parallel: true
 
       - name: System deps
         run: sudo apt-get update && sudo apt-get install -y gettext
-
-      - name: Install Poetry
-        run: |
-          python -m pip install --upgrade pip pipx
-          pipx install poetry
 
       - name: Install project (with dev tools)
         run: poetry install --with dev --no-interaction
