+++ .github/workflows/tests.yml
@@ -1,54 +1,150 @@
 name: Tests
 on: [push, pull_request]
-
 jobs:
-  tests:
-    name: ${{ matrix.test-name }} (py${{ matrix.python-version }})
+  test-DataExpansionFitter:
+    runs-on: ubuntu-latest
+    timeout-minutes: 60
+    strategy:
+      matrix:
+        python-version: ['3.9','3.10','3.11','3.12','3.13']
+    steps:
+      - uses: actions/checkout@v4
+      - name: Set up Python ${{ matrix.python-version }}
+        uses: actions/setup-python@v4
+        with:
+          python-version: ${{ matrix.python-version }}
+      - name: Setup
+        run: |
+          sudo apt-get update && sudo apt-get install -y gettext
+          python -m pip install --upgrade pip poetry
+          python -m poetry install --with dev --no-interaction
+      - name: Run tests
+        run: python -m poetry run pytest tests/test_DataExpansionFitter.py -v
+
+  test-EventTimesSampler:
+    runs-on: ubuntu-latest
+    timeout-minutes: 60
+    strategy:
+      matrix:
+        python-version: ['3.9','3.10','3.11','3.12','3.13']
+    steps:
+      - uses: actions/checkout@v4
+      - name: Set up Python ${{ matrix.python-version }}
+        uses: actions/setup-python@v4
+        with:
+          python-version: ${{ matrix.python-version }}
+      - name: Setup
+        run: |
+          sudo apt-get update && sudo apt-get install -y gettext
+          python -m pip install --upgrade pip poetry
+          python -m poetry install --with dev --no-interaction
+      - name: Run tests
+        run: python -m poetry run pytest tests/test_EventTimesSampler.py -v
+
+  test-TwoStagesFitter:
+    runs-on: ubuntu-latest
+    timeout-minutes: 60
+    strategy:
+      matrix:
+        python-version: ['3.9','3.10','3.11','3.12','3.13']
+    steps:
+      - uses: actions/checkout@v4
+      - name: Set up Python ${{ matrix.python-version }}
+        uses: actions/setup-python@v4
+        with:
+          python-version: ${{ matrix.python-version }}
+      - name: Setup
+        run: |
+          sudo apt-get update && sudo apt-get install -y gettext
+          python -m pip install --upgrade pip poetry
+          python -m poetry install --with dev --no-interaction
+      - name: Run tests
+        run: python -m poetry run pytest tests/test_TwoStagesFitter.py -v
+
+  test-TwoStagesFitterExact:
     runs-on: ubuntu-latest
     timeout-minutes: 120
     strategy:
-      fail-fast: false
       matrix:
-        python-version: ["3.9","3.10","3.11","3.12","3.13"]
-        test-name: [DataExpansionFitter, EventTimesSampler, TwoStagesFitter,
-                    TwoStagesFitterExact, screening, model_selection, remaining]
+        python-version: ['3.9','3.10','3.11','3.12','3.13']
     steps:
       - uses: actions/checkout@v4
-
-      - uses: actions/setup-python@v5
+      - name: Set up Python ${{ matrix.python-version }}
+        uses: actions/setup-python@v4
         with:
           python-version: ${{ matrix.python-version }}
-          cache: poetry            # cache Poetry deps keyed by poetry.lock
-
-      - uses: snok/install-poetry@v1
-        with:
-          version: latest
-          virtualenvs-create: true
-          virtualenvs-in-project: true
-          installer-parallel: true
-
-      - name: System deps
-        run: sudo apt-get update && sudo apt-get install -y gettext
-
-      - name: Install project (with dev tools)
-        run: poetry install --with dev --no-interaction
-
-      - name: Run selected tests
-        shell: bash
+      - name: Setup
         run: |
[…]
