### 2025-09-01 #8815 @tomer1812
Hi @rich2355,

I have expanded PyDTS compatibility across Python versions 3.9–3.13, with the necessary version constraints on numpy and scipy for each. The new release (v0.9.8) works seamlessly in Google Colab based on my testing. Please let me know if you encounter any issues.
In addition, I added a Dependencies section to both the README file and the documentation homepage.

Thank you!
Tomer
