### 2025-09-22 #8815 @bkrayfield
👋 Hello @rich2355, @LingfengLuo0510, @CamDavidsonPilon,

I’m checking in to see how your reviews are progressing.

@rich2355, I noticed you and the author had some back-and-forth on changes, but I’m not sure where things currently stand.

Could each of you provide a brief status update on your checklist?
