### 2025-09-24 #8815 @tomer1812
Hi @rich2355 ,

Thank you very much for your constructive comments. Please find my point-by-point responses below:

1. The EventTimesSampler handles Z1 that includes only 0s and 1s correctly according to the type returned by numpy. I have updated the figure (used only for visualization of the example data) to clearly reflect the binary nature of this variable (0.5 is no longer shown, as suggested).

2. I revised the sampling parameters to allow J=1 to be sampled (its occurrence probabilities were by mistake very low as it was). I now print both the totals of (J, X) and the first observations. Sampling T\C=d+1 is not due to inconsistency, I added an additional clarification on administrative censoring as follows:
"Note that with $d=7$ possible event times, observations with $X=8$ are only permitted when $J=0$ (censoring). Observations with outcome $(J, X) = (0, d+1)$ are included in the risk set throughout the entire follow-up period $t=1,...,d$, with no event observed.
Additionally, note that the columns $C$ and $T$ are not observed in practice and are included only for sampling purposes. The observed outcomes consist solely of the columns $J$ and $X$."

3. I added the following clarifications:
"The model summary table for $\beta_j$ reports the estimated parameters for each risk along with their standard errors. The model summary table for $\alpha_{jt}$ reports, for each $J$ and $X$, the number of observed events (column $n_{jt}$), the status of the optimization procedure (column success), and the estimated parameter $\alpha_{jt}$."

Prediction metrics can be accessed separately if desired. I added:
"Note that each of the above can be added either separately or for specific times/events if not all columns are required. For example, see partial functions such as predict_prob_events(), predict_hazard_all(), and others in the API documentation."

4. An introductory page has been added to the examples section of the documentation ( https://tomer1812.github.io/pydts/examples-intro/ ), with a note that all figures in pydts.example_utils are intended only for documentation purposes and are not generalizable to arbitrary datasets.

5. I removed the empty banners.

Thank you again for your thoughtful feedback, which has substantially improved both the clarity and reliability of the documentation.

Sincerely,
Tomer
