### 2025-09-25 #8815 @CamDavidsonPilon
Just a few notes (apologies if this overlaps with the other reviewers - I haven't read them yet):

1. The paper mentions a `mimic_df`, but this isn't part of the project. The output figure is lovely, but we can't reproduce it. Is `mimic_df` available somewhere?
2. I cloned the repo, and created a virtualenv, and then `pip install -e .`. All went well. However, running the following code found in your README in the `python` shell (i.e. not IPython) raised a `ModuleNotFoundError: No module named 'IPython'` error:

```
from pydts.fitters import TwoStagesFitter
from pydts.examples_utils.generate_simulations_data import generate_quick_start_df

# Generate a synthetic dataset with 10,000 patients,
# 5 covariates, 14 discrete time intervals, and 2 competing events
patients_df = generate_quick_start_df(n_patients=10000, n_cov=5, d_times=14, j_events=2, pid_col='pid', seed=0)

# Initialize and fit the discrete-time competing-risk model
fitter = TwoStagesFitter()
fitter.fit(df=patients_df.drop(['C', 'T'], axis=1))

# Display model summary
fitter.print_summary()
```
Should IPython be a dependency, or maybe a `try... except ImportError`? Note that I _didn't_ use `poetry` (I'm old school I guess) initially! Maybe this is intended!

3. Nitpick: the repo on github should have a homepage targeting your docs. That's the first place I look.
4. Nitpick: My preference would be to "hide" the `WORKERS = psutil.cpu_count(logical=False)` inside the class, so that at import time, it's not prematurely computed.
5. Nitpick: `.github/workflows/tests.yml` is a bit non-standard, why not just run the the single test suite with `poetry run pytest`?
6. Actually, I suggest adding a section in your readme (or docs) about _how_ to install and run the test suite. This would have helped me with my importerror above.
7. Test suite passes locally 👍
8. Some tests could be improved to check for valid outputs via `asserts`, and not just running them.
