### 2025-10-01 #8815 @tomer1812
Hi @CamDavidsonPilon and @rich2355 ,

Thank you for your great comments! I'll work on it and provide a point by point response in a few days.

Thanks,
Tomer
