### 2025-10-10 1e38a53
Documentation updates added predict_full v0.9.10
