+++ tests/test_DataExpansionFitter.py
@@ -122,6 +122,9 @@ class TestDataExpansionFitter(unittest.TestCase):
     def test_predict_cumulative_incident_function_case_successful_predict(self):
         self.fitted_model.predict_cumulative_incident_function(df=self.df.drop(['C', 'T'], axis=1))
 
+    def test_predict_full_case_successful_predict(self):
+        self.fitted_model.predict_full(df=self.df.drop(['C', 'T'], axis=1))
+
     def test_predict_hazard_jt_case_hazard_already_on_df(self):
         df_temp = get_real_hazard(self.df.drop(['C', 'T', 'X', 'J'], axis=1).set_index('pid').copy(),
                                   real_coef_dict=self.real_coef_dict,
