### 2025-10-10 bac4ad5
WORKERS calculation update
