+++ src/pydts/examples_utils/plots.py
@@ -6,11 +6,13 @@ from matplotlib import pyplot as plt
 import matplotlib.gridspec as gridspec
 import seaborn as sns
 from lifelines import KaplanMeierFitter
+from ..evaluation import *
 from ..config import *
 import os
 import string
 import warnings
 warnings.filterwarnings('ignore')
+from .mimic_consts import *
 
 
 def add_panel_text(ax, text, xplace=-0.15, fsz=17):
@@ -773,3 +775,284 @@ def plot_sampled_covariates_figure(observations_df, fname, show=True):
     fig.savefig(fname, dpi=300)
     if show:
         plt.show()
+
+
+def create_mimic_analysis_results(fitters_table, n_splits, penalizers, penalty_cv_search, chosen_eta,
+                                  lee_fitter, new_fitter, reg_fitter):
+    slicer = pd.IndexSlice
+    filename = 'mimic_summary_.png'
+    case = f'mimic_final_'
+
+    lee_alpha_ser = lee_fitter.get_alpha_df().loc[:, slicer[:, [COEF_COL, STDERR_COL]]].unstack().sort_index()
+    lee_beta_ser = lee_fitter.get_beta_SE().loc[:, slicer[:, [COEF_COL, STDERR_COL]]].unstack().sort_index()
+
+    two_step_alpha_k_results = new_fitter.alpha_df[['J', 'X', 'alpha_jt']]
+    two_step_beta_k_results = new_fitter.get_beta_SE().unstack().to_frame()
+
+    reg_two_step_alpha_k_results = reg_fitter.alpha_df[['J', 'X', 'alpha_jt']]
+    reg_two_step_beta_k_results = reg_fitter.get_beta_SE().unstack().to_frame()
+
+    lee_alpha_k_results = lee_alpha_ser.to_frame()
+    lee_beta_k_results = lee_beta_ser.to_frame()
+
+    # Cache results
+    two_step_alpha_k_results.to_csv(os.path.join(OUTPUT_DIR, f'{case}_two_step_alpha.csv'))
+    two_step_beta_k_results.to_csv(os.path.join(OUTPUT_DIR, f'{case}_two_step_beta.csv'))
+    reg_two_step_alpha_k_results.to_csv(os.path.join(OUTPUT_DIR, f'{case}_reg_two_step_alpha.csv'))
+    reg_two_step_beta_k_results.to_csv(os.path.join(OUTPUT_DIR, f'{case}_reg_two_step_beta.csv'))
+    lee_alpha_k_results.to_csv(os.path.join(OUTPUT_DIR, f'{case}_lee_alpha.csv'))
+    lee_beta_k_results.to_csv(os.path.join(OUTPUT_DIR, f'{case}_lee_beta.csv'))
+
+    covariates = [c for c in fitters_table.columns if c not in ['pid', 'J', 'X']]
+
+    two_step_alpha_k_results = pd.read_csv(os.path.join(OUTPUT_DIR, f'{case}_two_step_alpha.csv'),
+                                           index_col=['J', 'X'])
+    two_step_beta_k_results = pd.read_csv(os.path.join(OUTPUT_DIR, f'{case}_two_step_beta.csv'),
+                                          index_col=[0, 1])
+    reg_two_step_alpha_k_results = pd.read_csv(os.path.join(OUTPUT_DIR, f'{case}_reg_two_step_alpha.csv'),
+                                               index_col=['J', 'X'])
+    reg_two_step_beta_k_results = pd.read_csv(os.path.join(OUTPUT_DIR, f'{case}_reg_two_step_beta.csv'),
+                                              index_col=[0, 1])
+    lee_alpha_k_results = pd.read_csv(os.path.join(OUTPUT_DIR, f'{case}_lee_alpha.csv'),
+                                      index_col=[0, 1, 2])
[…]
