+++ src/pydts/examples_utils/datasets.py
@@ -1,8 +1,336 @@
-import pandas as pd
 from pydts.config import *
+from pydts.examples_utils.mimic_consts import *
+from pydts.evaluation import *
+from .mimic_consts import *
+import numpy as np
+import pandas as pd
+import os
+from sklearn.preprocessing import StandardScaler
+import importlib
+import subprocess
+import sys
+
+slicer = pd.IndexSlice
+
 
 DATASETS_DIR = os.path.join(os.path.dirname((os.path.dirname(__file__))), 'datasets')
+slicer = pd.IndexSlice
+
+
+def ensure_package(pkg_name):
+    try:
+        importlib.import_module(pkg_name)
+    except ImportError:
+        print(f"{pkg_name} not found. Installing...")
+        subprocess.check_call([sys.executable, "-m", "pip", "install", pkg_name])
+
 
 def load_LOS_simulated_data():
     os.path.join(os.path.dirname(__file__))
     return pd.read_csv(os.path.join(DATASETS_DIR, 'LOS_simulated_data.csv'))
+
+
+def get_mimic_df(mimic_data_dir, return_table1=False, chunksize=10 ** 6, minimum_year=2014):
+    """
+    Generate an example DataFrame from MIMIC-IV v2.0 files.
+
+    Args:
+        mimic_data_dir (str): Path to the directory containing the MIMIC-IV v2.0 data files.
+        print_table1_latex (bool, optional): If True, print a LaTeX-formatted version of Table 1.
+            Defaults to False.
+        chunksize (int, optional): Number of rows per chunk when iterating through large files.
+            Reduce this value if system memory is limited. Defaults to 1,000,000.
+        minimum_year (int, optional): Minimum year of admission to include in the dataset. Defaults to 2014.
+
+    Returns:
+        pd.DataFrame: A sample dataset extracted from the MIMIC-IV v2.0 files.
+    """
+
+    patients_file = os.path.join(mimic_data_dir, 'hosp', 'patients.csv.gz')
+    admissions_file = os.path.join(mimic_data_dir, 'hosp', 'admissions.csv.gz')
+    lab_file = os.path.join(mimic_data_dir, 'hosp', 'labevents.csv.gz')
+    lab_meta_file = os.path.join(mimic_data_dir, 'hosp', 'd_labitems.csv.gz')
+
+    patients_df = pd.read_csv(patients_file, compression='gzip')
+    COLUMNS_TO_DROP = ['dod']
+    patients_df.drop(COLUMNS_TO_DROP, axis=1, inplace=True)
+
+    admissions_df = pd.read_csv(admissions_file, compression='gzip', parse_dates=[ADMISSION_TIME_COL,
+                                                                                  DISCHARGE_TIME_COL,
+                                                                                  DEATH_TIME_COL,
+                                                                                  ED_REG_TIME, ED_OUT_TIME])
+
+    COLUMNS_TO_DROP = ['hospital_expire_flag', 'edouttime', 'edregtime', 'deathtime', 'language']
+    admissions_df.drop(COLUMNS_TO_DROP, axis=1, inplace=True)
+
+    admissions_df = admissions_df.merge(patients_df, on=[SUBJECT_ID_COL])
+
+    # # Calculate Age at Admission and Group of Admission Year
[…]
