### 2025-10-11 cfb8de2
get expanded df update v0.10.0
