### 2025-10-11 eb851f1
reverse change in get expanded df v0.10.1
