### 2025-10-11 baad32f
get expanded df fix v0.10.2
