### 2025-10-11 0b8395d
mimic dataset get dummies fix v0.10.3
