### 2025-10-19 #8815 @CamDavidsonPilon
I would like to see more complete test suite before accepting (see my note above on the tests). @tomer1812, I can help if you'd like.
