+++ tests/test_DataExpansionFitter.py
@@ -2,6 +2,7 @@ import unittest
 from src.pydts.examples_utils.generate_simulations_data import generate_quick_start_df
 from src.pydts.fitters import DataExpansionFitter
 import numpy as np
+import pandas as pd
 from src.pydts.utils import get_real_hazard
 
 
@@ -50,23 +51,57 @@ class TestDataExpansionFitter(unittest.TestCase):
             self.m.fit(df=self.df.drop(['C', 'T'], axis=1), covariates=['Z6'])
 
     def test_fit_case_correct_fit(self):
-        # Fit should be successful
+        # Fit should be successful and produce valid models
         m = DataExpansionFitter()
         m.fit(df=self.df.drop(['C', 'T'], axis=1))
+        
+        # Validate that fitting produced expected attributes
+        self.assertEqual(m.events, [1, 2])
+        self.assertEqual(m.times[:-1], list(range(1, 11, 1)))
+        self.assertEqual(m.covariates, ['Z1', 'Z2', 'Z3', 'Z4', 'Z5'])
+        self.assertIsNotNone(m.event_models)
+        coef1 = m.get_beta_SE()[(1, '   coef   ')].astype(float)
+        coef2 = m.get_beta_SE()[(2, '   coef   ')].astype(float)
+        self.assertTrue((coef1.min() > -1.6))
+        self.assertTrue((coef1.max() < 0.4))
+        self.assertTrue((coef2.min() > -1.7))
+        self.assertTrue((coef2.max() < 0.3))
+        coef1 = m.get_alpha_df()[(1, '   coef   ')].astype(float)
+        coef2 = m.get_alpha_df()[(2, '   coef   ')].astype(float)
+        self.assertTrue((coef1.min() > -2.3))
+        self.assertTrue((coef1.max() < 0))
+        self.assertTrue((coef2.min() > -3.7))
+        self.assertTrue((coef2.max() < -1.8))
 
     def test_fit_with_kwargs(self):
         import statsmodels.api as sm
         m = DataExpansionFitter()
         m.fit(df=self.df.drop(columns=['C', 'T']), models_kwargs=dict(family=sm.families.Binomial()))
+        
+        # Validate that fitting with kwargs produced expected attributes
+        self.assertIsNotNone(m.events)
+        self.assertIsNotNone(m.times)
+        self.assertIsNotNone(m.covariates)
+        self.assertEqual(len(m.events), 2)  # Should have 2 events
+        self.assertEqual(len(m.times[:-1]), 10)  # Should have 10 event time points
+        self.assertEqual(len(m.covariates), 5)  # Should have 5 covariates
 
     def test_print_summary(self):
         self.fitted_model.print_summary()
 
     def test_get_beta_SE(self):
-        self.fitted_model.get_beta_SE()
+        result = self.fitted_model.get_beta_SE()
+        
+        # Validate output structure and content
+        self.assertIsInstance(result, pd.DataFrame)
+        self.assertFalse(result.empty)
 
     def test_get_alpha_SE(self):
-        self.fitted_model.get_alpha_df()
+        result = self.fitted_model.get_alpha_df()
+        
+        # Validate output structure
+        self.assertIsInstance(result, pd.DataFrame)
+        self.assertFalse(result.empty)
 
     def test_predict_hazard_jt_case_covariate_not_in_df(self):
         # Covariates columns used in fit (here ['Z1','Z2','Z3','Z4','Z5']) must be passed in df to .predict()
@@ -89,24 +124,46 @@ class TestDataExpansionFitter(unittest.TestCase):
                 df=self.df.drop(['C', 'T'], axis=1), event=self.fitted_model.events[0], t=1000)
 
     def test_predict_hazard_jt_case_successful_predict(self):
-        self.fitted_model.predict_hazard_jt(
[…]

[…]
