+++ tests/test_EventTimesSampler.py
@@ -30,7 +30,22 @@ class TestEventTimesSampler(unittest.TestCase):
                 2: -np.log([1, 3, 4, 3, 2]),
             }
         }
-        ets.sample_event_times(patients_df, hazard_coefs=real_coef_dict, seed=seed)
+        result = ets.sample_event_times(patients_df, hazard_coefs=real_coef_dict, seed=seed)
+        
+        # Validate output structure
+        self.assertIsInstance(result, pd.DataFrame)
+        self.assertEqual(len(result), n_patients)
+        
+        # Check required columns exist
+        self.assertIn('T', result.columns)  # Event time
+        self.assertIn('J', result.columns)  # Event type
+        
+        # Validate event types are within expected range
+        self.assertTrue(result['J'].isin([0, 1, 2]).all())
+        
+        # Validate event times are positive and within reasonable range
+        self.assertTrue((result['T'] > 0).all())
+        self.assertTrue((result['T'] <= (ets.d_times+1)).all())
 
     def test_sample_3_events_4_covariates(self):
         n_cov = 4
@@ -40,7 +55,7 @@ class TestEventTimesSampler(unittest.TestCase):
         patients_df = pd.DataFrame(data=np.random.uniform(low=0.0, high=1.0, size=[n_patients, n_cov]),
                                    columns=covariates)
 
-        ets = EventTimesSampler(d_times=15, j_event_types=2)
+        ets = EventTimesSampler(d_times=15, j_event_types=3)  # Fixed: should be 3 events, not 2
         real_coef_dict = {
             "alpha": {
                 1: lambda t: -1 - 0.3 * np.log(t),
@@ -53,7 +68,22 @@ class TestEventTimesSampler(unittest.TestCase):
                 3: -np.log([0.5, 2, 3, 4]),
             }
         }
-        ets.sample_event_times(patients_df, hazard_coefs=real_coef_dict, seed=seed)
+        result = ets.sample_event_times(patients_df, hazard_coefs=real_coef_dict, seed=seed)
+        
+        # Validate output structure
+        self.assertIsInstance(result, pd.DataFrame)
+        self.assertEqual(len(result), n_patients)
+        
+        # Check required columns exist
+        self.assertIn('T', result.columns)  # Event time
+        self.assertIn('J', result.columns)  # Event type
+        
+        # Validate event types are within expected range (1, 2, 3)
+        self.assertTrue(result['J'].isin([0, 1, 2, 3]).all())
+        
+        # Validate event times are positive and within reasonable range
+        self.assertTrue((result['T'] > 0).all())
+        self.assertTrue((result['T'] <= (ets.d_times+1)).all())
 
     def test_sample_hazard_censoring(self):
         seed = 0
@@ -73,7 +103,20 @@ class TestEventTimesSampler(unittest.TestCase):
                 0: -np.log([0.8, 3, 3, 2.5, 2]),
             }
         }
-        ets.sample_hazard_lof_censoring(patients_df, censoring_coef_dict, seed)
+        result = ets.sample_hazard_lof_censoring(patients_df, censoring_coef_dict, seed)
+        
+        # Validate hazard censoring was applied
+        self.assertIsInstance(result, pd.DataFrame)
+        self.assertEqual(len(result), n_patients)
+        self.assertIn('C', result.columns)  # Censoring time
+        
+        # Check censoring times are valid
+        self.assertTrue((result['C'] > 0).all())
+        self.assertTrue((result['C'] <= (ets.d_times+1)).all())
+        
+        # Check that original covariates are preserved
+        for cov in covariates:
+            self.assertIn(cov, result.columns)
 
     def test_sample_independent_censoring(self):
         n_cov = 4
@@ -84,7 +127,16 @@ class TestEventTimesSampler(unittest.TestCase):
 
[…]
