+++ tests/test_evaluation.py
@@ -55,7 +55,12 @@ class TestEvaluation(unittest.TestCase):
         fitter.fit(df=self.patients_df.drop(['C', 'T'], axis=1), nb_workers=1)
         pred_df = fitter.predict_prob_event_j_all(self.patients_df, event=1)
         esauc = event_specific_auc_at_t(pred_df, event=1, t=9)
-        print(esauc)
+        
+        # Validate AUC properties
+        self.assertIsInstance(esauc, (float, np.float64))
+        self.assertGreaterEqual(esauc, 0)
+        self.assertLessEqual(esauc, 1)
+        self.assertTrue(np.isfinite(esauc))
 
     def test_event_specific_auc_perfect_model(self):
         cov_df = self.patients_df[self.covariates]
@@ -68,7 +73,12 @@ class TestEvaluation(unittest.TestCase):
                                              for t in range(1, self.ets.d_times+1)])
         pred_df = pd.concat([self.patients_df, probs_j_at_t], axis=1)
         esauc = event_specific_auc_at_t(pred_df, event=event, t=9)
-        print(esauc)
+        
+        # Perfect model should have AUC better than random (0.5)
+        self.assertIsInstance(esauc, (float, np.float64))
+        self.assertGreater(esauc, 0.5)  # Should be better than random
+        self.assertLessEqual(esauc, 1.0)
+        self.assertTrue(np.isfinite(esauc))
 
     def test_event_specific_brier_score_perfect_model(self):
         cov_df = self.patients_df[self.covariates]
@@ -80,8 +90,13 @@ class TestEvaluation(unittest.TestCase):
                                     columns=[f'prob_j{event}_at_t{t}'
                                              for t in range(1, self.ets.d_times+1)])
         pred_df = pd.concat([self.patients_df, probs_j_at_t], axis=1)
-        esauc = event_specific_brier_score_at_t(pred_df, event=event, t=9)
-        print(esauc)
+        esbs = event_specific_brier_score_at_t(pred_df, event=event, t=9)
+        
+        # Perfect model should have reasonable Brier score
+        self.assertIsInstance(esbs, (float, np.float64))
+        self.assertGreaterEqual(esbs, 0)
+        self.assertLess(esbs, 0.25)  # Should be reasonable for perfect model
+        self.assertTrue(np.isfinite(esbs))
 
     def test_event_specific_auc_all_perfect_model(self):
         cov_df = self.patients_df[self.covariates]
@@ -94,7 +109,19 @@ class TestEvaluation(unittest.TestCase):
                                              for t in range(1, self.ets.d_times+1)])
         pred_df = pd.concat([self.patients_df, probs_j_at_t], axis=1)
         esauc = event_specific_auc_at_t_all(pred_df, event=event)
-        print(esauc)
+        
+        # Validate AUC all times results for perfect model
+        self.assertIsInstance(esauc, (pd.DataFrame, pd.Series, dict))
+        if isinstance(esauc, pd.Series):
+            # Should have valid AUC values
+            self.assertTrue((esauc >= 0.0).all())  # Valid AUC range
+            self.assertTrue((esauc <= 1.0).all())
+            self.assertTrue(esauc.notna().all())
+        elif isinstance(esauc, pd.DataFrame):
+            numeric_cols = esauc.select_dtypes(include=[np.number]).columns
+            for col in numeric_cols:
+                self.assertTrue((esauc[col] >= 0.0).all())  # Valid AUC range
+                self.assertTrue((esauc[col] <= 1.0).all())
 
     def test_event_specific_brier_score_all_perfect_model(self):
         cov_df = self.patients_df[self.covariates]
@@ -107,7 +134,19 @@ class TestEvaluation(unittest.TestCase):
[…]
