### 2025-10-21 #8815 @tomer1812
Dear @CamDavidsonPilon ,

Thank you for the tests suite important comment! I have updated the tests suite files, here's a summary of the main additions:

1. Expected attributes existance
2. Probabilities and metrics in the expected range [0,1]
3. Estimated coefficients of the test examples are in reasonable expected range
4. Data quality - no unexpected missing\infinite values
5. Cross-validation output structure
6. Screening - output validation of chosen covariates, threshold, and permutation results

Is there anything I may have overlooked that you believe would be worth adding?

Thank you,
Tomer
