### 2025-11-01 #8815 @tomer1812
Dear CamDavidsonPilon and rich2355,
Thank you very much for your thorough review and insightful comments!

Dear @bkrayfield,
I would be grateful if you could clarify the next steps in the process.

Thank you again to all of you for your time and consideration,
Tomer
