### 2025-11-12 #8815 @bkrayfield
@tomer1812,

Great job on the writing, I have no further comments there. Please complete the following so this can continue to move forward.

1. Update Zenodo to the current version in Github, and update the DOIs.
2. Correct and update the current reference list to include all DOIs. You can check your references with "@editorialbot check references."
3. Tell me what version number you would like to appear.

Once you complete these tasks, I can continue to move forward.

Thank you!
