### 2025-11-12 8c450cd
Update paper.bib
