+++ paper.bib
@@ -1,56 +1,3 @@
-@article{Polsterl:2020,
-  author  = {Sebastian Pölsterl},
-  title   = {scikit-survival: A Library for Time-to-Event Analysis Built on Top of scikit-learn},
-  journal = {Journal of Machine Learning Research},
-  year    = {2020},
-  volume  = {21},
-  number  = {212},
-  pages   = {1-6},
-  url     = {http://jmlr.org/papers/v21/20-729.html}
-}
-
-@article{tibshirani1996regression,
-  title={Regression Shrinkage and Selection via the Lasso},
-  author={Tibshirani, Robert},
-  journal={Journal of the Royal Statistical Society: Series B (Methodological)},
-  volume={58},
-  number={1},
-  pages={267--288},
-  year={1996},
-  url={https://www.jstor.org/stable/2346178}
-}
-@article{hoel1970ridge,
-  title={Ridge Regression: Biased Estimation for Nonorthogonal Problem.},
-  author={Hoel, Arthur E and Kennard, Robert W},
-  journal={Technometrics},
-  volume={42},
-  number={1},
-  pages={80--86},
-  year={1970},
-  doi={10.2307/1271436}
-}
-@article{zou2005regularization,
-  title={Regularization and Variable Selection via the Elastic Net},
-  author={Zou, Hui and Hastie, Trevor},
-  journal={Journal of the Royal Statistical Society: Series B (Statistical Methodology)},
-  volume={67},
-  number={2},
-  pages={301--320},
-  year={2005},
-  url = {https://www.jstor.org/stable/3647580}
-}
-
-@article{andersen1970asymptotic,
-  title={Asymptotic properties of conditional maximum-likelihood estimators},
-  author={Andersen, Erling Bernhard},
-  journal={Journal of the Royal Statistical Society: Series B (Methodological)},
-  volume={32},
-  number={2},
-  pages={283--301},
-  year={1970},
-  publisher={John Wiley \& Sons }
-}
-
 @article{zhao2012principled,
   title={Principled sure independence screening for Cox models with ultra-high-dimensional covariates},
   author={Zhao, Sihai Dave and Li, Yi},
@@ -63,73 +10,27 @@
   doi={10.1016/j.jmva.2011.08.002}
 }
 
-@article{prentice1978retrospective,
-  title={Retrospective studies and failure time models},
-  author={Prentice, Ross L and Breslow, Norman E},
-  journal={Biometrika},
-  volume={65},
-  number={1},
-  pages={153--158},
-  year={1978},
-  publisher={Oxford University Press}
+@article{Davidson-Pilon:2019,
+  doi = {10.21105/joss.01317},
+  url = {https://doi.org/10.21105/joss.01317},
+  year = {2019},
+  publisher = {The Open Journal},
+  volume = {4},
+  number = {40},
+  pages = {1317},
+  author = {Cameron Davidson-Pilon},
+  title = {lifelines: survival analysis in Python},
+  journal = {Journal of Open Source Software}
 }
 
-@article{schmid2021competing,
-  title={Competing Risks Analysis for Discrete Time-to-Event Data},
-  author={Schmid, Matthias and Berger, Moritz},
-  journal={WIREs Computational Statistics},
-  volume={13},
-  number={5},
-  pages={e1529},
-  year={2021},
-  doi={10.1002/wics.1529},
-}
-
-@book{klein_survival_2003,
-	title = {Survival Analysis},
-	publisher = {Springer-Verlag},
-	author = {Klein, John P. and Moeschberger, Melvin L.},
-	date = {2003},
[…]
