### 2025-11-12 0e957af
Update paper.bib
