+++ paper.bib
@@ -71,8 +71,7 @@
   volume  = {21},
   number  = {212},
   pages   = {1-6},
-  doi     = {https://dl.acm.org/doi/10.5555/3455716.3455928},
-  url     = {http://jmlr.org/papers/v21/20-729.html}
+  url     = {https://dl.acm.org/doi/10.5555/3455716.3455928},
 }
 
 @article{lee_analysis_2018,
