+++ README.md
@@ -2,7 +2,7 @@
 [![Tests](https://github.com/tomer1812/pydts/workflows/Tests/badge.svg)](https://github.com/tomer1812/pydts/actions?workflow=Tests)
 [![documentation](https://img.shields.io/badge/docs-mkdocs%20material-blue.svg?style=flat)](https://tomer1812.github.io/pydts)
 [![codecov](https://codecov.io/gh/tomer1812/pydts/branch/main/graph/badge.svg)](https://codecov.io/gh/tomer1812/pydts)
-[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.16991039.svg)](https://doi.org/10.5281/zenodo.16991039)
+[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.17593777.svg)](https://doi.org/10.5281/zenodo.17593777)
 
 # Discrete Time Survival Analysis  
 
