### 2025-11-18 5ee27c3
Update paper.bib
