+++ paper.bib
@@ -19,7 +19,7 @@
   number = {40},
   pages = {1317},
   author = {Cameron Davidson-Pilon},
-  title = {lifelines: survival analysis in Python},
+  title = {lifelines: survival analysis in {P}ython},
   journal = {Journal of Open Source Software}
 }
 
