+++ README.md
@@ -4,7 +4,7 @@
 [![codecov](https://codecov.io/gh/tomer1812/pydts/branch/main/graph/badge.svg)](https://codecov.io/gh/tomer1812/pydts)
 [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.17593777.svg)](https://doi.org/10.5281/zenodo.17593777)
 
-# Discrete Time Survival Analysis  
+# PyDTS: A Python Package for Discrete-Time Survival Analysis with Competing Risks and Optional Penalization
 
 *PyDTS* is a Python package designed for discrete-time survival analysis with competing risks, offering tools for model fitting, evaluation, and simulation.
 
