### 2025-11-19 df8d83b
Create CITATION.cff
