### 2025-11-30 f027e00
Update intro.md
