### 2025-11-30 a91b7dd
Update methodsintro.md
