### 2025-11-30 31c6584
Update methods.md
