+++ jss_replication.py
@@ -1,0 +1,1094 @@
+###################################################
+#  PyDTS: A Python Package for Discrete-Time Survival Analysis with Competing Risks (2022)
+#  Meir, Tomer*, Gutman, Rom*, and Gorfine, Malka.
+###################################################
+# Replication file
+
+
+from time import time
+from pydts.data_generation import EventTimesSampler
+from pydts.examples_utils.plots import add_panel_text, plot_events_occurrence, plot_example_pred_output, \
+    plot_models_coefficients, plot_times, compare_beta_models_for_example, plot_jss_reps_coef_std, \
+    plot_sampled_covariates_figure
+from pydts.examples_utils.generate_simulations_data import generate_quick_start_df
+from pydts.fitters import DataExpansionFitter, TwoStagesFitter, repetitive_fitters
+from pydts.model_selection import PenaltyGridSearch
+from pydts.cross_validation import PenaltyGridSearchCV, TwoStagesCV
+from pydts.examples_utils.mimic_consts import *
+from pydts.evaluation import *
+from tableone import TableOne
+import numpy as np
+import pandas as pd
+from matplotlib import pyplot as plt
+import warnings
+import os
+from sklearn.model_selection import train_test_split
+from sklearn.preprocessing import StandardScaler
+
+slicer = pd.IndexSlice
+
+pd.set_option("display.max_rows", 500)
+warnings.filterwarnings('ignore')
+nb_start = time()
+
+
+# The figures and tables of the paper were generated with lean_version = False.
+# Estimated total running time (Apple MacBook Pro 32Gb RAM):
+# lean_version = True: 3.5 hours
+# lean_version = False: 11 hours without MIMIC, 17.5 hours with MIMIC
+lean_version = False
+
+# Optional settings:
+# 1. Provide the MIMIC-IV v2.0 data directory (mimic-iv-2.0) to replicate the use-case section results.
+# Leaving mimic_data_dir = None skips the replication of Section 5.
+# For example:  mimic_data_dir = '/Users/tomer/data/mimic-iv-2.0'
+# The MIMIC dataset is accessible at https://physionet.org/content/mimiciv/2.0/ and subjected to PhysioNet credentials.
+mimic_data_dir = None
+
+# 2. Figures\tables output directory
+OUTPUT_DIR = ''
+
+# 3. Show figures using plt.show():
+PYPLOT_SHOW = True
+
+
+
+
+# Section 3: Simulating discrete time survival data with competing events
+############################################################################
+
+ets = EventTimesSampler(d_times=7, j_event_types=2)
+n_observations = 10000
+np.random.seed(0)
+
+observations_df = pd.DataFrame(columns=['Z1', 'Z2', 'Z3'])
+observations_df['Z1'] = np.random.binomial(n=1, p=0.5, size=n_observations)
+Z1_zero_index = observations_df.loc[observations_df['Z1'] == 0].index
+observations_df.loc[Z1_zero_index, 'Z2'] = \
+    np.random.normal(loc=72, scale=12, size=n_observations-observations_df['Z1'].sum())
+Z1_one_index = observations_df.loc[observations_df['Z1'] == 1].index
+observations_df.loc[Z1_one_index, 'Z2'] = \
[…]
